src/main/java/executor/service/step_execution/StepExecutionServiceSleep.java
package executor.service.step_execution;

import executor.model.Step;
import org.openqa.selenium.WebDriver;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.concurrent.TimeUnit;

@Service
public class StepExecutionServiceSleep implements StepExecutable {

    @Autowired
    public StepExecutionServiceSleep() {  }

    @Override
    public String getStepAction() {
        return "sleep";
    }

    @Override
    public void step(WebDriver webDriver, Step step) {
        try {
            TimeUnit.SECONDS.sleep(Long.parseLong(step.getValue()));
        } catch (InterruptedException e) {
            e.printStackTrace();
        }
    }
}
